refactor: reorganize Platform service into clean architecture
