--- PlatformService/Data/AppDbContext.cs
@@ -1,15 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-using PlatformService.Models;
-
-namespace PlatformService.Data
-{
-    public class AppDbContext: DbContext
-    {
-        public AppDbContext(DbContextOptions<AppDbContext> options) : base(options)
-        {
-        }
-
-        public DbSet<Platform> Platforms { get; set; }
-    }
-    
-}
