--- PlatformService/Data/PrepDb.cs
@@ -1,47 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-using PlatformService.Models;
-
-namespace PlatformService.Data
-{
-    public static class PrepDb
-    {
-        public static void PrepPopulation(IApplicationBuilder app, bool isProd)
-        {
-            using (var serviceScope = app.ApplicationServices.CreateScope())
-            {
-                SeedData(serviceScope.ServiceProvider.GetService<AppDbContext>(), isProd);
-            }
-        }
-
-        private static void SeedData(AppDbContext? context, bool isProd)
-        {
-            if (context == null) return;
-            if (isProd)
-            {
-                Console.WriteLine("--> Attempting to apply migrations...");
-                try
-                {
-                    context.Database.Migrate();
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine($"--> Could not apply migrations: {ex.Message}");
-                }
-            }
-            if (!context.Platforms.Any())
-            {
-                Console.WriteLine("--> Seeding data...");
-                context.Platforms.AddRange(
-                    new Platform { Name = "Dotnet", Publisher = "Microsoft", Cost = "Free" },
-                    new Platform { Name = "SQL Server Express", Publisher = "Microsoft", Cost = "Free" },
-                    new Platform { Name = "Kubernetes", Publisher = "Cloud Native Computing Foundation", Cost = "Free" }
-                );
-                context.SaveChanges();
-            }
-            else
-            {
-                Console.WriteLine("--> We already have data");
-            }
-        }
-    }
-}
