--- PlatformService/Data/Repos/Interfaces/IPlatformRepo.cs
@@ -1,13 +1,0 @@
-﻿using PlatformService.Models;
-
-namespace PlatformService.Data.Repos.Interfaces
-{
-    public interface IPlatformRepo
-    {
-        bool SaveChanges();
-
-        IEnumerable<Platform> GetAllPlatforms();
-        Platform? GetPlatformById(int id);
-        void CreatePlatform(Platform platform);
-    }
-}
