--- PlatformService/Data/Repos/PlatformRepo.cs
@@ -1,40 +1,0 @@
-﻿using PlatformService.Data.Repos.Interfaces;
-using PlatformService.Models;
-
-namespace PlatformService.Data.Repos
-{
-    public class PlatformRepo : IPlatformRepo
-    {
-        private readonly AppDbContext _context;
-
-        public PlatformRepo(AppDbContext context)
-        {
-            _context = context;
-        }
-
-        public void CreatePlatform(Platform platform)
-        {
-            if (string.IsNullOrWhiteSpace(platform.Name))
-            {
-                throw new ArgumentException("Platform name cannot be null or empty.", nameof(platform));
-            }
-
-            _context.Platforms.Add(platform);
-        }
-
-        public IEnumerable<Platform> GetAllPlatforms()
-        {
-            return _context.Platforms.OrderBy(p => p.Id).ToList();
-        }
-
-        public Platform? GetPlatformById(int id)
-        {
-            return _context.Platforms.FirstOrDefault(p => p.Id == id);
-        }
-
-        public bool SaveChanges()
-        {
-            return _context.SaveChanges() >= 0;
-        }
-    }
-}
