--- PlatformService/Dtos/PlatformCreateDto.cs
@@ -1,30 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace PlatformService.Dtos
-{
-    public class PlatformCreateDto
-    {
-        [Required]
-        public string Name { get; set; }
-
-        [Required]
-        public string Publisher { get; set; }
-
-        [Required]
-        public string Cost { get; set; }
-
-        public PlatformCreateDto()
-        {
-            Name = string.Empty;
-            Publisher = string.Empty;
-            Cost = string.Empty;
-        }
-
-        public PlatformCreateDto(string name, string publisher, string cost)
-        {
-            Name = name;
-            Publisher = publisher;
-            Cost = cost;
-        }
-    }
-}
