--- PlatformService/Models/Platform.cs
@@ -1,20 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace PlatformService.Models
-{
-    public class Platform
-    {
-        [Key]
-        [Required]
-        public int Id { get; set; }
-
-        [Required]
-        public required string Name { get; set; }
-
-        [Required]
-        public required string Publisher { get; set; }
-
-        [Required]
-        public required string Cost { get; set; }
-    }
-}
