--- PlatformService/PlatformService.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.14.36310.24 d17.14
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PlatformService", "PlatformService.csproj", "{AFFCB9BC-11F1-4403-EE41-48117AABFABA}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{AFFCB9BC-11F1-4403-EE41-48117AABFABA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AFFCB9BC-11F1-4403-EE41-48117AABFABA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AFFCB9BC-11F1-4403-EE41-48117AABFABA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AFFCB9BC-11F1-4403-EE41-48117AABFABA}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {34634114-074D-467A-A055-0AC5D32A2F55}
-	EndGlobalSection
-EndGlobal
