--- PlatformService/Profiles/PlatformsProfile.cs
@@ -1,12 +1,0 @@
-﻿namespace PlatformService.Profiles
-{
-    public class PlatformsProfile : AutoMapper.Profile
-    {
-        public PlatformsProfile()
-        {
-            // Source -> Target
-            CreateMap<Models.Platform, Dtos.PlatformReadDto>();
-            CreateMap<Dtos.PlatformCreateDto, Models.Platform>();
-        }
-    }
-}
