--- PlatformService/Startup.cs
@@ -1,85 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.HttpsPolicy;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
-using Microsoft.OpenApi.Models;
-using PlatformService.Data;
-using PlatformService.Data.Repos;
-using PlatformService.Data.Repos.Interfaces;
-using PlatformService.Services.Interfaces;
-using PlatformService.SyncDataServices.Http;
-
-namespace PlatformService
-{
-    public class Startup
-    {
-        public IConfiguration Configuration { get; }
-        private readonly IWebHostEnvironment _env;
-
-        public Startup(IConfiguration configuration, IWebHostEnvironment env)
-        {
-            Configuration = configuration;
-            _env = env;
-        }
-
-        public void ConfigureServices(IServiceCollection services)
-        {
-            if (_env.IsProduction())
-            {
-                Console.WriteLine("--> Using SQL Server Db");
-                services.AddDbContext<Data.AppDbContext>(options =>
-                    options.UseSqlServer(Configuration.GetConnectionString("PlatformsConnection")));
-            }
-            else
-            {
-                Console.WriteLine("--> Using InMem Db");
-                services.AddDbContext<Data.AppDbContext>(options =>
-                    options.UseInMemoryDatabase("InMem"));
-            }
-            
-            services.AddScoped<IPlatformRepo, PlatformRepo>();
-            services.AddScoped<IPlatformService, Services.PlatformService>();
-            services.AddHttpClient<ICommandDataClient, HttpCommandDataClient>();
-            services.AddControllers();
-            services.AddAutoMapper(AppDomain.CurrentDomain.GetAssemblies());
-            services.AddSwaggerGen(c =>
-            {
-                c.SwaggerDoc("v1", new OpenApiInfo { Title = "PlatformService", Version = "v1" });
-            });
-
-            Console.WriteLine($"--> CommandService Endpoint {Configuration["CommandService"]}");
-        }
-
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
-        {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-                app.UseSwagger();
-                app.UseSwaggerUI(c => c.SwaggerEndpoint("/swagger/v1/swagger.json", "PlatformService v1"));
-            }
-
-            // app.UseHttpsRedirection();
-
-            app.UseRouting();
-
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapControllers();
-            });
-
-            PrepDb.PrepPopulation(app, env.IsProduction());
-        }
-    }
-}
