--- PlatformService/SyncDataServices/Http/ICommandDataClient.cs
@@ -1,9 +1,0 @@
-﻿using PlatformService.Dtos;
-
-namespace PlatformService.SyncDataServices.Http
-{
-    public interface ICommandDataClient
-    {
-        Task SendPlatformToCommand(PlatformReadDto platform);
-    }
-}
