+++ src/Platform/Platform.Api/Properties/launchSettings.json
