+++ src/Platform/Platform.Api/Startup.cs
@@ -1,0 +1,42 @@
+using Company.Platform.Application;
+using Company.Platform.Infrastructure;
+using Company.Platform.Infrastructure.Data;
+
+namespace Company.Platform.Api;
+
+public class Startup
+{
+    public IConfiguration Configuration { get; }
+
+    public Startup(IConfiguration configuration)
+    {
+        Configuration = configuration;
+    }
+
+    public void ConfigureServices(IServiceCollection services)
+    {
+        services.AddApplication();
+        services.AddInfrastructure(Configuration);
+        services.AddControllers();
+        services.AddEndpointsApiExplorer();
+        services.AddSwaggerGen();
+    }
+
+    public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+    {
+        if (env.IsDevelopment())
+        {
+            app.UseSwagger();
+            app.UseSwaggerUI();
+        }
+
+        app.UseRouting();
+
+        app.UseEndpoints(endpoints =>
+        {
+            endpoints.MapControllers();
+        });
+
+        PrepDb.PrepPopulation(app, env.IsProduction());
+    }
+}
