+++ src/Platform/Platform.Api/appsettings.Development.json
