+++ src/Platform/Platform.Api/appsettings.json
