+++ src/Platform/Platform.Application/Repos/IPlatformRepo.cs
@@ -1,0 +1,12 @@
+using Company.Platform.Domain.Models;
+
+namespace Company.Platform.Application.Repos;
+
+public interface IPlatformRepo
+{
+    bool SaveChanges();
+
+    IEnumerable<Platform> GetAllPlatforms();
+    Platform? GetPlatformById(int id);
+    void CreatePlatform(Platform platform);
+}
