+++ src/Platform/Platform.Application/Services/Interfaces/IPlatformService.cs
@@ -1,0 +1,10 @@
+using Company.Platform.Application.Dtos;
+
+namespace Company.Platform.Application.Services.Interfaces;
+
+public interface IPlatformService
+{
+    IEnumerable<PlatformReadDto> GetAllPlatforms();
+    PlatformReadDto? GetPlatformById(int id);
+    Task<PlatformReadDto?> CreatePlatformAsync(PlatformCreateDto platformCreateDto);
+}
