+++ src/Platform/Platform.Application/SyncDataServices/Http/ICommandDataClient.cs
@@ -1,0 +1,8 @@
+using Company.Platform.Application.Dtos;
+
+namespace Company.Platform.Application.SyncDataServices.Http;
+
+public interface ICommandDataClient
+{
+    Task SendPlatformToCommand(PlatformReadDto platform);
+}
