+++ src/Platform/Platform.Domain/Models/Platform.cs
@@ -1,0 +1,20 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Company.Platform.Domain.Models
+{
+    public class Platform
+    {
+        [Key]
+        [Required]
+        public int Id { get; set; }
+
+        [Required]
+        public required string Name { get; set; }
+
+        [Required]
+        public required string Publisher { get; set; }
+
+        [Required]
+        public required string Cost { get; set; }
+    }
+}
