+++ src/Platform/Platform.Infrastructure/Data/AppDbContext.cs
@@ -1,0 +1,13 @@
+using Company.Platform.Domain.Models;
+using Microsoft.EntityFrameworkCore;
+
+namespace Company.Platform.Infrastructure.Data;
+
+public class AppDbContext : DbContext
+{
+    public AppDbContext(DbContextOptions<AppDbContext> options) : base(options)
+    {
+    }
+
+    public DbSet<Platform> Platforms => Set<Platform>();
+}
