+++ src/Platform/tests/Platform.IntegrationTests/SmokeTests.cs
@@ -1,0 +1,24 @@
+using Company.Platform.Api;
+using Microsoft.AspNetCore.Mvc.Testing;
+using System.Threading.Tasks;
+using Xunit;
+
+namespace Company.Platform.IntegrationTests;
+
+public class SmokeTests : IClassFixture<WebApplicationFactory<Program>>
+{
+    private readonly WebApplicationFactory<Program> _factory;
+
+    public SmokeTests(WebApplicationFactory<Program> factory)
+    {
+        _factory = factory;
+    }
+
+    [Fact]
+    public async Task GetPlatforms_EndpointReturnsSuccess()
+    {
+        var client = _factory.CreateClient();
+        var response = await client.GetAsync("/api/platforms");
+        Assert.True(response.IsSuccessStatusCode || response.StatusCode == System.Net.HttpStatusCode.NotFound);
+    }
+}
