+++ src/Platform/tests/Platform.UnitTests/SmokeTests.cs
@@ -1,0 +1,12 @@
+using Xunit;
+
+namespace Company.Platform.UnitTests;
+
+public class SmokeTests
+{
+    [Fact]
+    public void TrueIsTrue()
+    {
+        Assert.True(true);
+    }
+}
